Reduce memory footprint of searches
Result search group values are now stored in a central index
so that identical values are only stored once and retreived
by reference. Also reduces the size of the result object send
back to the main process via results queue which greatly
reduces the time it takes the send/receive it.

We now use the shelve library to manage the MPCache underlying
key value store which reduces complexity and improves
performance (although there are more improvements that can be
done but will addresses in seperate patches).