Implement more efficient file indexing
The constraints engine indexes files
to get start of line pointers used in the binary
search. This is an expensive task for large files so
we ensure that we hold as little as possible in
memory and try to obtain line positions as fast as
possible. This is achieved with the addition of a
new MPCacheSharded class and faster line ending
lookup with reduced reliance on tell/seek.