Fixes case where search paths overlap
If the paths used to register searches overlap once
expanded, they will cause the same file to be searched
concurrently which breaks the MPCache and is also
superfluous. This patch fixes that problem and applies
some minor optimisations to the way we extract
datetime from start of line to apply constraints.

Also fixes support for applying constraints to files
containing unicode characters by ensuring that we escape
rather than decode those charaters.