Optimise data transfer between principle and worker procs
SearchResultMinimal is used to transder results between
the worker processes and the main collector. As a result
it must be a small as possible in order to keep transfer
fast and memory footprint low. There were several
unnecessary variables being stored and duplicated in this
object that have now been removed maing the transfer
faster and using less memory. Transfers are also batched
so as to reduce interruption on searches.

Also removes unnecssary use of multiprocessing.Queue for
single thread usecase.