Pre-allocate result store blocks
Rather than query the parent result store for
every new allocation (which is very slow) we
speed things up by giving each worker a
pre-allocated block that it syncs to the main
store once finished. This does mean that
deduplication is no longer global and done per
worker but it does speed things up a lot.