Don't fail on MaxSearchableLineLengthReached
When searchkit looks for line feeds in a file, it will raise the
MaxSearchableLineLengthReached after reading 1048576 chars and not
finding one.
Now we catch the exception, log a message and skip searching the
rest of the file. This approach allows for successful completion
of the parallel searches.

Little refactoring in 'apply_to_file' was needed to avoid pylint
'too-many-return-statements / R0911' complaint.

Closes-issue: https://github.com/dosaboy/searchkit/issues/21
Signed-off-by: Marcin Wilk <marcin.wilk@canonical.com>